feat: 커스텀 구분자 지정 및 구조 변경
Numbers 객체 추가
